feat: 🎸 Added community detection (clustering) on front
✅ Closes: https://trello.com/c/7ZPjwGq8/64-front-identificar-comunidades-de-usuarios-clustering

+++ wasm-lib/src/social_network/output_node.rs
@@ -7,7 +7,7 @@ pub struct OutputNode {
     pub id: String,
     pub name: String,
     pub val: usize,
-    pub group: u8,
+    pub group: u32,
 }
 
 impl OutputNode {
@@ -16,7 +16,7 @@ impl OutputNode {
             id: node.id.clone(),
             name: node.name.clone(),
             val: 0,
-            group: node.group.clone() as u8,
+            group: node.group.clone() as u32,
         }
     }
 }
